fix: edge_stencil burn_at WGSL i32/u32 mismatch
idx (i32 from signed multiply) used with >> u32 and & u32 operators;
WGSL requires matching element types for bit ops on storage arrays.
Cast idx to u32: let idx = u32(r * u.ncols + c); mask 31u not 31.

+++ gigamercado/_vulkan.py
@@ -99,8 +99,8 @@ _WGSL["edge"] = """struct UBO { ncols: i32, xmin: i32, ymin: i32, nrows: i32 };
 const QP = 0.78539816339744830962;
 
 fn burn_at(r: i32, c: i32) -> bool {
-    let idx = r * u.ncols + c;
-    return (burn[idx >> 5u] >> (idx & 31)) & 1u != 0u;
+    let idx = u32(r * u.ncols + c);
+    return (burn[idx >> 5u] >> (idx & 31u)) & 1u != 0u;
 }
 
 @compute @workgroup_size(16, 16)
